refactor(remediate): extract the configured sequencing loop (module size) + use the internal-push helper
The self-guardrail blocked cli.ts at the large-file threshold; the pure
sequencing loop + its git seams move to configured-loop.ts (the plan-cli
precedent), re-exported from cli.ts. The resume marker push now builds
its argv via internalGitPushArgs (the arch gate's own catch — a machine
push must carry --no-verify).
